Add Auth.js TRUST_HOST config
Running built app, this has to be set to avoid UntrustedHost error:
https://github.com/nextauthjs/next-auth/issues/8327#issuecomment-2051648141

Note any value (including `"false"`) in the env will be resolved as
true, here set it to empty, user has the responsibility to change it.

https://authjs.dev/reference/core#trusthost

Also not the vars has to appear in wrangler config. NextJS env doesn't
exposing to next-on-pages.