Allow control initial timeline scroll by keyboard
When first time enter Facts page, user can't use keyboard PageDown and
so on to do scrolling. That is because initial focus is taken by <body>.

Now add a hidden handle <a> in the beginning of first timeline, and try
focusing it as a solution.

The reason introducing another <a> instead of just focus the timeline
body is to avoid focus ring style appear.