Fix highlight missing when restore to 2 characters
Problem: when backspace edit keyword, e.g., from `fo` to `f`, the filter
becomes invalid because the length of keyword is less than 2. Then when
input back to `fo`, it should highlight `fo`, but it doesn't.

The cause is that the KeywordHighlighter still caches the `fo` as
previous keyword. We have missed to reset the cache.