Standout the sticky header of list if has scrolled
For pick "index" and "my" listing panels, the header part is sticky at
top, and can be visually messed up with the underlying content.

Now detect the scroll position, if moved, add some outline and box
shadow to the header, make it clearer.

Also add a handler when click the header, scroll the content to top.