Avoid Tooltip Menu hanging when ref element moved
When click an item of PicksPanel layout menu, the panel will be moved by
animation (to the selected layout), while the menu is still floating at
the original position, which can distract users.

Now extend TooltipContentMenu to add a `dismissOnClick` prop, when set,
any click on it will set the Tooltip to close.