Fix newly created Pick has missing props
The new props was fetched from backend result. Only returns the
`returning` of db insert is not enough, for example it doesn't have
`userName` set.

Now do an extra db query to fetch the new record again.

Also fix an error that trying to format `createdAt` while it is null.