Fix Date error when save pick and go to pick list
After creating a new draft Pick, then go to "public" pick list, an
invalidDate error will occur.

The problem is the `createdAt` was replaced with an `invalidDate` to
hide it from user. And the PickRow can't format such a bad date.

The `invalidDate` implementation was once introduced but later consider
bad practice, removed in 3027840e. The `refreshPickAtom` is a part
missing to be removed, the bomb just exploded.