Hint clicking book icon to start reading pick
In small screen, previous hint tells user to click the box to see pick
timeline. Which lacks to care a case that if user has not select any
pick yet, there is no timeline to follow.

Now add a check if a pick is selected, display different hint for that.