Display date with leading zero removed
We are going to have new Fact data with date value like `789` instead of
common `1998-01-01` style, however `789` will be sorted AFTER `1998`
which is not what we want.

Thus we make the data `0775` to cover such rare case.

Here just adjust the display only, remove the leading zero to make it
looks right.